:ukraine: Update image paths and model structure, enhance homepage design
Several changes were made including updating the image paths in the AppFixtures file. Also, "Trait" namespace has been corrected in both the Product model and SlugTrait file. The homepage design of the application was improved by adding a display for each category. Furthermore, the database configuration in .env.local was adjusted, and the Turbo-core status in controllers.json was disabled. Finally, the database query in MainController was updated to fetch all categories whose parent value is null.