:ukraine: Rearrange homepage content and introduce category route
The homepage content originally in _nav.html.twig was moved into index.html.twig for better organization and streamlined page composition. As part of the same modification, the content in _nav.html.twig was cleaned up. Additionally, a new 'categories' route was added to the MainController for the rendering of individual category pages. Lastly, a new categories.html.twig file was created as a template for these category pages.