:ukraine: Add product detail view and update product presentation layout
Implemented a new detail view for individual products and made adjustments to the layout of the product presentation in the main view. This enhances the user experience by providing more detailed product information in a cleaner, more organized format. The name of the product image file was also corrected in the fixture data.