:ukraine: Add user registration functionality
This commit includes the creation of a new User registration form, the linking of the User entity to the form for validation and the creation of a new RegistrationController to handle the registration logic. The form including all necessary fields is displayed on the register.html.twig page. Also, the login page was updated to provide navigation to the registration page.