:ukraine: Update navigation bar based on user state
This commit changes the display of the navigation links depending on the user's status. If the user is logged in, it shows the logout and register buttons. If the user is not logged in, it shows a login button. A success alert for logged-in users was also moved from the login page to the index page.