:ukraine: Refine admin interface and update dependencies
In the admin interface, unnecessary edit and delete buttons were removed and sections were centered for a cleaner look. In addition, dependencies in 'composer.lock' were updated, which was prompted by a shift in the database connection in '.env.local'. Changes in 'templates/admin/product/index.html.twig' also enhanced the display of product information.