:ukraine: Add Cart and CartItem entities, update Product and User entities
Added new Cart and CartItem classes under the Entity namespace, which include fields and methods to manage a shopping cart and items within it. Included relevant annotations for ORM and API Platform as well. Also, updated the existing User and Product entities to establish relationships with Cart and CartItem. Specifically, each User has a Cart and each Product can be in multiple CartItems.