:ukraine: Add contact page with form and link in navigation bar
A new 'contact.html.twig' template file has been created with a form for users to get in touch. Additionally, the link to the contact page has been added to the navigation bar in '_nav.html.twig'. A corresponding function 'contact' has also been added to the 'MainController.php' in order to render the contact page.